+++ projects/js-packages/eslint-changed/bin/eslint-changed.js
@@ -9,7 +9,7 @@ const fs = require( 'fs' );
 const path = require( 'path' );
 const chalk = require( 'chalk' );
 
-const APP_VERSION = '2.0.2';
+const APP_VERSION = '2.0.3-alpha';
 
 const { program } = require( 'commander' );
 program