+++ projects/packages/wordads/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.1] - 2022-04-26
+### Changed
+- Update package.json metadata.
+
 ## 0.1.0 - 2022-04-19
 ### Added
 - Added WordAds package for new UI work
@@ -12,3 +16,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - PHPCS: Fix `WordPress.Security.ValidatedSanitizedInput`
 - Updated package dependencies.
+
+[0.1.1]: https://github.com/Automattic/jetpack-wordads/compare/v0.1.0...v0.1.1